api/requirements.txt
aniso8601==7.0.0
click==7.1.2
Flask==1.1.2
Flask-GraphQL==2.0.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
promise==2.3
Rx==1.6.1
six==1.15.0
Werkzeug==1.0.1
flask-cors
requests
python-dotenv
python-dateutil
elasticsearch7